+++ misc/dbt-materialize/dbt/adapters/materialize/relation.py
@@ -40,6 +40,11 @@ class MaterializeRelationType(StrEnum):
 class MaterializeRelation(PostgresRelation):
     type: Optional[MaterializeRelationType] = None
 
+    # Materialize does not have a 63-character limit for relation
+    # names, unlike PostgreSQL (see dbt-core #2727)
+    def relation_max_name_length(self):
+        return 16384
+
     @classproperty
     def get_relation_type(cls) -> Type[MaterializeRelationType]:
         return MaterializeRelationType